chore: add confirmations when terminating sessions and revoking tokens

+++ framework/core/js/src/forum/components/AccessTokensList.tsx
@@ -146,12 +146,15 @@ export default class AccessTokensList<CustomAttrs extends IAccessTokensListAttrs
   }
 
   revoke(token: AccessToken) {
+    if (!confirm(extractText(app.translator.trans('core.forum.security.revoke_access_token_confirmation')))) return;
+
     this.loading[token.id()!] = true;
 
     return token.delete().then(() => {
       this.loading[token.id()!] = false;
       this.attrs.ondelete && this.attrs.ondelete(token);
-      app.alerts.show({ type: 'success' }, app.translator.trans('core.forum.security.session_terminated', { count: 1 }));
+      const key = this.attrs.type === 'session' ? 'session_terminated' : 'token_revoked';
+      app.alerts.show({ type: 'success' }, app.translator.trans(`core.forum.security.${key}`, { count: 1 }));
       m.redraw();
     });
   }
